oci-py/requirements.txt
oci
InquirerPy
python-dotenv